feat(svelte5): nav bar
