+++ src/lib/index.ts
@@ -12,8 +12,7 @@ export { default as IconTrashCan } from './IconTrashCan.svelte';
 export { default as ImageArticle } from './ImageArticle.svelte';
 export { default as IrocoLogo } from './IrocoLogo.svelte';
 // export { default as Loader } from './Loader.svelte';
-// export { default as NavBar } from './NavBar.svelte';
-// export { default as Navigation } from './Navigation.svelte';
+export { default as NavBar } from './NavBar.svelte';
 export { default as NumberInput } from './NumberInput.svelte';
 export { default as RadioButton } from './RadioButton.svelte';
 export { default as TextInput } from './TextInput.svelte';
